small error handling additions in status function

+++ farmer.js
@@ -104,6 +104,7 @@ farmer.get("/update", async (req, res) => {
 
 				try {
 					let status = await create_main_log_object(item.root_folder);
+					if (!status || !status.length) return resolve();
 					let stat = status.map(function(folder) {
 						return new Promise(function(end, stop) {
 							connection.query("DELETE FROM status WHERE farmer_id=?", item.id, (err) => {
